refactor new file download functions to new file
